refactor : Menu 내 menuProducts 필드 일급컬렉션 구현
