+++ src/main/java/kitchenpos/domain/Menu.java
@@ -1,11 +1,9 @@
 package kitchenpos.domain;
 
 import java.math.BigDecimal;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
-import javax.persistence.CascadeType;
+import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
@@ -13,9 +11,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-
-import kitchenpos.exception.WrongPriceException;
 
 @Entity
 public class Menu {
@@ -30,8 +25,8 @@ public class Menu {
 	@JoinColumn(name = "menu_group_id")
 	private MenuGroup menuGroup;
 
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "menu", cascade = CascadeType.ALL, orphanRemoval = true)
-	private List<MenuProduct> menuProducts = new ArrayList<>();
+	@Embedded
+	private MenuProducts menuProducts = new MenuProducts();
 
 	protected Menu() {
 	}
@@ -48,21 +43,7 @@ public class Menu {
 	}
 
 	private void addMenuProducts(BigDecimal price, List<Product> products, List<Long> quantities) {
-		BigDecimal sum = BigDecimal.ZERO;
-		for (int i = 0; i < products.size(); i++) {
-			this.menuProducts.add(MenuProduct.create(this, products.get(i), quantities.get(i)));
-			sum = sum.add(products.get(i).getPrice().multiply(BigDecimal.valueOf(quantities.get(i))));
-		}
-		validatePrice(price, sum);
-	}
-
-	private void validatePrice(BigDecimal price, BigDecimal sum) {
-		if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
-			throw new WrongPriceException("메뉴의 가격이 없거나 0보다 작습니다.");
-		}
-		if (price.compareTo(sum) > 0) {
-			throw new WrongPriceException("메뉴의 가격이 상품가격의 총합보다 클 수 없습니다.");
-		}
+		this.menuProducts.add(this, price, products, quantities);
 	}
 
 	public Long getId() {
@@ -82,6 +63,6 @@ public class Menu {
 	}
 
 	public List<MenuProduct> getMenuProducts() {
-		return menuProducts;
+		return menuProducts.getMenuProducts();
 	}
 }
